factor out queue for postgres and sqlite too
