+++ common/persistence/sql/sqlplugin/sqlite/task_queues.go
@@ -39,103 +39,16 @@ const (
 
 	lockTaskQueueQry = `SELECT range_id FROM task_queues ` +
 		`WHERE range_hash = ? AND task_queue_id = ?`
-	// *** Task_Queues Table Above ***
-
-	// *** Tasks Below ***
-	getTaskMinMaxQry = `SELECT task_id, data, data_encoding ` +
-		`FROM tasks ` +
-		`WHERE range_hash = ? AND task_queue_id = ? AND task_id >= ? AND task_id < ? ` +
-		` ORDER BY task_id LIMIT ?`
-
-	getTaskMinQry = `SELECT task_id, data, data_encoding ` +
-		`FROM tasks ` +
-		`WHERE range_hash = ? AND task_queue_id = ? AND task_id >= ? ORDER BY task_id LIMIT ?`
-
-	createTaskQry = `INSERT INTO ` +
-		`tasks(range_hash, task_queue_id, task_id, data, data_encoding) ` +
-		`VALUES(:range_hash, :task_queue_id, :task_id, :data, :data_encoding)`
-
-	// deleteTaskQry = `DELETE FROM tasks ` +
-	// 	`WHERE range_hash = ? AND task_queue_id = ? AND task_id = ?`
-
-	rangeDeleteTaskQry = `DELETE FROM tasks ` +
-		`WHERE range_hash = ? AND task_queue_id = ? AND task_id IN (SELECT task_id FROM
-		 tasks WHERE range_hash = ? AND task_queue_id = ? AND task_id < ? ` +
-		`ORDER BY task_queue_id,task_id LIMIT ? ) `
 )
-
-// InsertIntoTasks inserts one or more rows into tasks table
-func (mdb *db) InsertIntoTasks(
-	ctx context.Context,
-	rows []sqlplugin.TasksRow,
-) (sql.Result, error) {
-	return mdb.conn.NamedExecContext(ctx,
-		createTaskQry,
-		rows,
-	)
-}
-
-// SelectFromTasks reads one or more rows from tasks table
-func (mdb *db) SelectFromTasks(
-	ctx context.Context,
-	filter sqlplugin.TasksFilter,
-) ([]sqlplugin.TasksRow, error) {
-	var err error
-	var rows []sqlplugin.TasksRow
-	switch {
-	case filter.ExclusiveMaxTaskID != nil:
-		err = mdb.conn.SelectContext(ctx,
-			&rows, getTaskMinMaxQry,
-			filter.RangeHash,
-			filter.TaskQueueID,
-			*filter.InclusiveMinTaskID,
-			*filter.ExclusiveMaxTaskID,
-			*filter.PageSize,
-		)
-	default:
-		err = mdb.conn.SelectContext(ctx,
-			&rows, getTaskMinQry,
-			filter.RangeHash,
-			filter.TaskQueueID,
-			*filter.ExclusiveMaxTaskID,
-			*filter.PageSize,
-		)
-	}
-	if err != nil {
-		return nil, err
-	}
-	return rows, nil
-}
-
-// DeleteFromTasks deletes one or more rows from tasks table
-func (mdb *db) DeleteFromTasks(
-	ctx context.Context,
-	filter sqlplugin.TasksFilter,
-) (sql.Result, error) {
-	if filter.ExclusiveMaxTaskID == nil {
-		return nil, serviceerror.NewInternal("missing ExclusiveMaxTaskID parameter")
-	}
-	if filter.Limit == nil || *filter.Limit == 0 {
-		return nil, serviceerror.NewInternal("missing limit parameter")
-	}
-	return mdb.conn.ExecContext(ctx,
-		rangeDeleteTaskQry,
-		filter.RangeHash,
-		filter.TaskQueueID,
-		filter.RangeHash,
-		filter.TaskQueueID,
-		*filter.ExclusiveMaxTaskID,
-		*filter.Limit,
-	)
-}
 
 // InsertIntoTaskQueues inserts one or more rows into task_queues table
 func (mdb *db) InsertIntoTaskQueues(
 	ctx context.Context,
 	row *sqlplugin.TaskQueuesRow,
+	v sqlplugin.MatchingTaskVersion,
 ) (sql.Result, error) {
 	return mdb.conn.NamedExecContext(ctx,
-		createTaskQueueQry,
+		sqlplugin.SwitchTaskQueuesTable(createTaskQueueQry, v),
 		row,
 	)
 }
@@ -144,9 +57,10 @@ func (mdb *db) InsertIntoTaskQueues(
 func (mdb *db) UpdateTaskQueues(
 	ctx context.Context,
 	row *sqlplugin.TaskQueuesRow,
+	v sqlplugin.MatchingTaskVersion,
 ) (sql.Result, error) {
 	return mdb.conn.NamedExecContext(ctx,
-		updateTaskQueueQry,
+		sqlplugin.SwitchTaskQueuesTable(updateTaskQueueQry, v),
 		row,
 	)
 }
@@ -155,20 +69,21 @@ func (mdb *db) UpdateTaskQueues(
 func (mdb *db) SelectFromTaskQueues(
 	ctx context.Context,
 	filter sqlplugin.TaskQueuesFilter,
+	v sqlplugin.MatchingTaskVersion,
 ) ([]sqlplugin.TaskQueuesRow, error) {
 	switch {
 	case filter.TaskQueueID != nil:
 		if filter.RangeHashLessThanEqualTo != 0 || filter.RangeHashGreaterThanEqualTo != 0 {
 			return nil, serviceerror.NewInternal("range of hashes not supported for specific selection")
 		}
-		return mdb.selectFromTaskQueues(ctx, filter)
+		return mdb.selectFromTaskQueues(ctx, filter, v)
 	case filter.RangeHashLessThanEqualTo != 0 && filter.PageSize != nil:
 		if filter.RangeHashLessThanEqualTo < filter.RangeHashGreaterThanEqualTo {
 			return nil, serviceerror.NewInternal("range of hashes bound is invalid")
 		}
-		return mdb.rangeSelectFromTaskQueues(ctx, filter)
+		return mdb.rangeSelectFromTaskQueues(ctx, filter, v)
 	case filter.TaskQueueIDGreaterThan != nil && filter.PageSize != nil:
-		return mdb.rangeSelectFromTaskQueues(ctx, filter)
+		return mdb.rangeSelectFromTaskQueues(ctx, filter, v)
 	default:
 		return nil, serviceerror.NewInternal("invalid set of query filter params")
 	}
@@ -177,12 +92,13 @@ func (mdb *db) SelectFromTaskQueues(
 func (mdb *db) selectFromTaskQueues(
 	ctx context.Context,
 	filter sqlplugin.TaskQueuesFilter,
+	v sqlplugin.MatchingTaskVersion,
 ) ([]sqlplugin.TaskQueuesRow, error) {
 	var err error
 	var row sqlplugin.TaskQueuesRow
 	err = mdb.conn.GetContext(ctx,
 		&row,
-		getTaskQueueQry,
+		sqlplugin.SwitchTaskQueuesTable(getTaskQueueQry, v),
 		filter.RangeHash,
 		filter.TaskQueueID,
 	)
@@ -195,6 +111,7 @@ func (mdb *db) selectFromTaskQueues(
 func (mdb *db) rangeSelectFromTaskQueues(
 	ctx context.Context,
 	filter sqlplugin.TaskQueuesFilter,
+	v sqlplugin.MatchingTaskVersion,
 ) ([]sqlplugin.TaskQueuesRow, error) {
 	var err error
 	var rows []sqlplugin.TaskQueuesRow
@@ -202,7 +119,7 @@ func (mdb *db) rangeSelectFromTaskQueues(
 	if filter.RangeHashLessThanEqualTo != 0 {
 		err = mdb.conn.SelectContext(ctx,
 			&rows,
-			listTaskQueueWithHashRangeQry,
+			sqlplugin.SwitchTaskQueuesTable(listTaskQueueWithHashRangeQry, v),
 			filter.RangeHashGreaterThanEqualTo,
 			filter.RangeHashLessThanEqualTo,
 			filter.TaskQueueIDGreaterThan,
@@ -211,7 +128,7 @@ func (mdb *db) rangeSelectFromTaskQueues(
 	} else {
 		err = mdb.conn.SelectContext(ctx,
 			&rows,
-			listTaskQueueQry,
+			sqlplugin.SwitchTaskQueuesTable(listTaskQueueQry, v),
 			filter.RangeHash,
 			filter.TaskQueueIDGreaterThan,
 			*filter.PageSize,
@@ -227,9 +144,10 @@ func (mdb *db) rangeSelectFromTaskQueues(
 func (mdb *db) DeleteFromTaskQueues(
 	ctx context.Context,
 	filter sqlplugin.TaskQueuesFilter,
+	v sqlplugin.MatchingTaskVersion,
 ) (sql.Result, error) {
 	return mdb.conn.ExecContext(ctx,
-		deleteTaskQueueQry,
+		sqlplugin.SwitchTaskQueuesTable(deleteTaskQueueQry, v),
 		filter.RangeHash,
 		filter.TaskQueueID,
 		*filter.RangeID,
@@ -240,11 +158,12 @@ func (mdb *db) DeleteFromTaskQueues(
 func (mdb *db) LockTaskQueues(
 	ctx context.Context,
 	filter sqlplugin.TaskQueuesFilter,
+	v sqlplugin.MatchingTaskVersion,
 ) (int64, error) {
 	var rangeID int64
 	err := mdb.conn.GetContext(ctx,
 		&rangeID,
-		lockTaskQueueQry,
+		sqlplugin.SwitchTaskQueuesTable(lockTaskQueueQry, v),
 		filter.RangeHash,
 		filter.TaskQueueID,
 	)
